Make folder path clickable in detailed view
